pyproject.toml
[project]
name = "langchain-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gh>=0.0.4",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.3",
    "pylatexenc>=2.10",
    "python-dotenv>=1.1.1",
]
